添付ファイル（インライン可能）・マルチパートメールへの対応

+++ lib/queued_mail/deliver_email_job.rb
@@ -4,8 +4,9 @@ module QueuedMail
     
     def self.perform(args)
       message = QueuedMail::Message.find(args["message_id"])
+      return unless message
+
       message.lock!
-      
       QueuedMail::Mailer.original_email(message).deliver
       message.destroy
     end
